[Bugfix] fix links to other pages in docs (#245)
* do nothing

* update links to be prefixed with /docs/ in markdown files

* use /docs/*.md for document routes

* consistent spacing

* revert links in markdown docs

* use the use_signal hook to clean up some code

* remove /docs prefix on dioxus routes
