+++ frontend/src/components.rs
@@ -10,6 +10,7 @@ mod event_teams_list;
 mod league_penalty_types_table;
 mod league_registration_buttons;
 mod penalty_display;
+mod markdown;
 mod team_token_input;
 
 pub use ass_entry::AssEntry;
@@ -22,4 +23,5 @@ pub use event_teams_list::EventTeamsList;
 pub use league_penalty_types_table::LeaguePenaltyTypesTable;
 pub use league_registration_buttons::LeagueRegistrationButtons;
 pub use penalty_display::PenaltyDisplay;
+pub use markdown::Markdown;
 pub use team_token_input::{all_tokens_known, resolve_value_to_team_ids, TeamSelectionField, TeamTokenInput};
