+++ frontend/src/pages/arctos_schedule_script.rs
@@ -1,26 +1,9 @@
-use crate::api;
+use crate::components::Markdown;
 use dioxus::prelude::*;
 
 #[component]
 pub fn ArctosScheduleScript() -> Element {
-    let data = use_resource(|| async move {
-        api::markdown_page("arctos-schedule-script")
-            .await
-            .map_err(|e| e.to_string())
-    });
-    let val = data.value();
-
     rsx! {
-        if let Some(Ok(d)) = val.read().as_ref() {
-            div { class: "row",
-                div { class: "col-lg-10 mx-auto",
-                    div { dangerous_inner_html: "{d.html}" }
-                }
-            }
-        } else if let Some(Err(e)) = val.read().as_ref() {
-            p { class: "text-danger", "{e}" }
-        } else {
-            p { "Loading…" }
-        }
+        Markdown { page: "arctos-schedule-script" }
     }
 }
